Move DeltaSetIndexMap to variations
